+++ Prism.Game/include/ECS/Components/ResourceSpawnComponent.h
@@ -6,12 +6,12 @@ namespace ECS {
 	namespace Components {
 		struct ResourceSpawnComponent : Component {
 			ResourceSpawnComponent() = default;
-			ResourceSpawnComponent(int gatherRate_, std::string type_) : 
+			ResourceSpawnComponent(float gatherRate_, std::string type_) : 
 				gatherRate(gatherRate_), resourceType(type_){};
 
 			std::string resourceType{};
 
-			int gatherRate = 0;
+			float gatherRate = 0;
 			Component* Clone() override {
 				return new ResourceSpawnComponent(gatherRate, resourceType);
 			}
